cleanup events and mutation interface a bit more
Mutations can be applied at any level (including top-level multiple-model client)
by calling apply(eventName, index, change)

eventName must be `localChange` or `remoteChange`.  change is a single change or
an array of changes.  index refers to number of levels within change.Path that
has already been processed. This allows the same change object to be passed down
instead of creating a new change object with the truncated path.

The app can also directly mutate a model.  Mutating a model via `apply` results
in its event bubbling up the corresponding event with data set to:

`{change, index, before, after}`

where change is the applied change, and as before `index` refers to the slice of
change.Path that describes where the change happened.  The `before` and `after`
fields indicate the before and after values of the current model being changed.

Note that `localChange` events bubble up (but not remote change events as these
naturally are applied top-down) with the change.Path correspondingly getting
prepended to.  In particular, by the time the event bubbles up to the client
object, it contains the modelID as the first element.

The `before` and `after` fields are good for managing persistent datastructures.
Note that the client and ModelManager (the top level objects) are not really
persistent.  Also, the before/after for a client `localChange` event conveniently
stores the affected ModelManager rather than the singleton client.

It is possible to access a sub-path from any level by calling `getValue(path)`.
In particular, `client.getValue([modelID])` returns the value of the top-level
model. (It does not return the model manager)

These changes are poorly tested as the test infrastructure for testing these is
slowly being built up now.