move initialization from SyncBridge directly to Transport
It makes more sense for the transport to initialize an uncached
model from the network and create a bridge for it and present it
to the app layer.

It avoids the unnecessary initializing check and general ugliness
of creating a model + bridge if all one wants to do is use the
transport to work with a model.