db: cleanup append() interface
The append() interface did not allow returning a "modified" change. This is needed for derivations where the user-applied change is either not valid at all or should be modified slightly.

All the append implementations are not quite right at this point of mulitple derivations (i.e. A derives B and C; and both B & C are used to derive D; the change pushed from D will land on say B but its reflection on C will not be available yet and when it is fetched, it may not be properly applied). The implementations may actually work but more testing is needed to validate these types of issues.